==> 0g_prepare_seasonal_climate_mode_data.ipynb <==
# Get climate mode data for desired months

from dask_jobqueue import PBSCluster
from dask.distributed import Client

# One node on Gadi has 48 cores - try and use up a full core before going to multiple nodes (jobs)

walltime = '00:20:00'
cores = 2
memory = '8GB'

cluster = PBSCluster(walltime=str(walltime), cores=cores, memory=str(memory),
                     job_extra=['-l ncpus='+str(cores),
                                '-l mem='+str(memory),
                                '-P xv83',
                                '-l storage=gdata/xv83+gdata/rt52+scratch/xv83'],
                     header_skip=["select"])

cluster.scale(jobs=1)
client = Client(cluster)

client

import xarray as xr
import numpy as np
import pandas as pd

import functions as fn

%load_ext autoreload
%autoreload 2

# Load coffee data

growing_calendar = pd.read_csv('/g/data/xv83/dr6273/work/projects/coffee/data/coffee_country_growing_calendar_extended.csv',
                               index_col=0)
growing_calendar.head()

# Process climate modes

sst_dataset = 'hadisst'

### Nino3.4

nino34 = xr.open_zarr('/g/data/xv83/dr6273/work/data/'+sst_dataset+'/climate_modes/'+sst_dataset+'_nino34_1979-2020.zarr',
                      consolidated=True)

fn.process_and_write(ds=nino34,
                     dataset=sst_dataset,
                     var='nino34',
                     event_list=[['Flowering', 'both_tails', '1_std'],
                                 ['Growing', 'both_tails', '1_std']],
                     detrend=False,
                     coffee_df=growing_calendar,
                     spatial_field=False)

fn.process_and_write(ds=nino34,
                     dataset=sst_dataset,
                     var='nino34_detrended',
                     event_list=[['Flowering', 'both_tails', '1_std'],
                                 ['Growing', 'both_tails', '1_std']],
                     detrend=False,
                     coffee_df=growing_calendar,
                     spatial_field=False)

### DMI

dmi = xr.open_zarr('/g/data/xv83/dr6273/work/data/'+sst_dataset+'/climate_modes/'+sst_dataset+'_dmi_1979-2020.zarr',
                   consolidated=True)

fn.process_and_write(ds=dmi,
                     dataset=sst_dataset,
                     var='dmi',
                     event_list=[['Flowering', 'both_tails', '1_std'],
                                 ['Growing', 'both_tails', '1_std']],
                     detrend=False,
                     coffee_df=growing_calendar,
                     spatial_field=False)

fn.process_and_write(ds=dmi,
                     dataset=sst_dataset,
                     var='dmi_detrended',
                     event_list=[['Flowering', 'both_tails', '1_std'],
                                 ['Growing', 'both_tails', '1_std']],
                     detrend=False,
                     coffee_df=growing_calendar,
                     spatial_field=False)

### Atlantic Nino

atl_nino = xr.open_zarr('/g/data/xv83/dr6273/work/data/'+sst_dataset+'/climate_modes/'+sst_dataset+'_atl_nino_1979-2020.zarr',
                        consolidated=True)

fn.process_and_write(ds=atl_nino,
                     dataset=sst_dataset,
                     var='atl_nino',
                     event_list=[['Flowering', 'both_tails', '1_std'],
                                 ['Growing', 'both_tails', '1_std']],
                     detrend=False,
                     coffee_df=growing_calendar,
                     spatial_field=False)

fn.process_and_write(ds=atl_nino,
                     dataset=sst_dataset,
                     var='atl_nino_detrended',
                     event_list=[['Flowering', 'both_tails', '1_std'],
                                 ['Growing', 'both_tails', '1_std']],
                     detrend=False,
                     coffee_df=growing_calendar,
                     spatial_field=False)

### TNA

tna = xr.open_zarr('/g/data/xv83/dr6273/work/data/'+sst_dataset+'/climate_modes/'+sst_dataset+'_tna_1979-2020.zarr',
                   consolidated=True)

fn.process_and_write(ds=tna,
                     dataset=sst_dataset,
                     var='tna',
                     event_list=[['Flowering', 'both_tails', '1_std'],
                                 ['Growing', 'both_tails', '1_std']],
                     detrend=False,
                     coffee_df=growing_calendar,
                     spatial_field=False)

fn.process_and_write(ds=tna,
                     dataset=sst_dataset,
                     var='tna_detrended',
                     event_list=[['Flowering', 'both_tails', '1_std'],
                                 ['Growing', 'both_tails', '1_std']],
                     detrend=False,
                     coffee_df=growing_calendar,
                     spatial_field=False)

### TSA

tsa = xr.open_zarr('/g/data/xv83/dr6273/work/data/'+sst_dataset+'/climate_modes/'+sst_dataset+'_tsa_1979-2020.zarr',
                   consolidated=True)

fn.process_and_write(ds=tsa,
                     dataset=sst_dataset,
                     var='tsa',
                     event_list=[['Flowering', 'both_tails', '1_std'],
                                 ['Growing', 'both_tails', '1_std']],
                     detrend=False,
                     coffee_df=growing_calendar,
                     spatial_field=False)

fn.process_and_write(ds=tsa,
                     dataset=sst_dataset,
                     var='tsa_detrended',
                     event_list=[['Flowering', 'both_tails', '1_std'],
                                 ['Growing', 'both_tails', '1_std']],
                     detrend=False,
                     coffee_df=growing_calendar,
                     spatial_field=False)

### MJO

Compute these statistics:
- Number of days in each phase per coffee season (i.e. growing/flowering)
- Average magnitude during each phase over season

mjo_anoms = xr.open_zarr('/g/data/xv83/dr6273/work/data/era5/climate_modes/mjo_monthly_anoms_1979-2020.zarr',
                       consolidated=True)

mjo_anoms_dt = xr.open_zarr('/g/data/xv83/dr6273/work/data/era5/climate_modes/mjo_monthly_anoms_1979-2020_detrended.zarr',
                          consolidated=True)

fn.process_and_write(ds=mjo_anoms,
                     dataset='era5',
                     var='mjo_days_per_month',
                     event_list=[['Flowering', 'both_tails', '1_std'],
                                 ['Growing', 'both_tails', '1_std']],
                     detrend=False,
                     coffee_df=growing_calendar,
                     spatial_field=False)

fn.process_and_write(ds=mjo_anoms,
                     dataset='era5',
                     var='mjo_mean_amplitude',
                     event_list=[['Flowering', 'both_tails', '1_std'],
                                 ['Growing', 'both_tails', '1_std']],
                     detrend=False,
                     coffee_df=growing_calendar,
                     spatial_field=False)

fn.process_and_write(ds=mjo_anoms_dt,
                     dataset='era5',
                     var='mjo_days_per_month_detrended',
                     event_list=[['Flowering', 'both_tails', '1_std'],
                                 ['Growing', 'both_tails', '1_std']],
                     detrend=False,
                     coffee_df=growing_calendar,
                     spatial_field=False)

fn.process_and_write(ds=mjo_anoms_dt,
                     dataset='era5',
                     var='mjo_mean_amplitude_detrended',
                     event_list=[['Flowering', 'both_tails', '1_std'],
                                 ['Growing', 'both_tails', '1_std']],
                     detrend=False,
                     coffee_df=growing_calendar,
                     spatial_field=False)

# Close cluster

client.close()
cluster.close()
